requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
apache-beam==2.69.0
asn1crypto==1.5.1
attrs==25.4.0
backports.tarfile==1.2.0
beartype==0.21.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cloud-sql-python-connector==1.18.5
crcmod==1.7
cryptography==46.0.3
dnspython==2.8.0
docopt==0.6.2
docstring_parser==0.17.0
fastavro==1.12.1
fasteners==0.20
frozenlist==1.8.0
google-api-core==2.28.1
google-apitools==0.5.31
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-aiplatform==1.127.0
google-cloud-bigquery==3.38.0
google-cloud-bigquery-storage==2.34.0
google-cloud-bigtable==2.34.0
google-cloud-core==2.5.0
google-cloud-datastore==2.21.0
google-cloud-dlp==3.33.0
google-cloud-language==2.18.0
google-cloud-pubsub==2.33.0
google-cloud-pubsublite==1.12.0
google-cloud-recommendations-ai==0.10.18
google-cloud-resource-manager==1.15.0
google-cloud-secret-manager==2.25.0
google-cloud-spanner==3.59.0
google-cloud-storage==2.19.0
google-cloud-videointelligence==2.17.0
google-cloud-vision==3.11.0
google-crc32c==1.7.1
google-genai==1.50.0
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.65.5
grpcio-status==1.65.5
h11==0.16.0
hdfs==2.7.3
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jsonpickle==3.4.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==25.6.0
keyrings.google-artifactregistry-auth==1.1.2
more-itertools==10.8.0
multidict==6.7.0
numpy==2.2.6
oauth2client==4.1.3
objsize==0.7.1
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
overrides==7.7.0
packaging==25.0
pg8000==1.31.5
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==18.1.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydot==1.4.2
PyJWT==2.10.1
pymongo==4.15.4
PyMySQL==1.1.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-tds==1.17.1
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
redis==5.3.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.28.0
rsa==4.9.1
scramp==1.4.6
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
